Let there be light

+++ src/entitylist.cuh
@@ -16,7 +16,7 @@ public:
 __device__ bool EntityList::hit(const Ray& r, float tmin, float tmax, HitRecord& rec) const {
     HitRecord temp_rec;
     bool hit_anything = false;
-    double closest_so_far = tmax;
+    float closest_so_far = tmax;
 
     for (int i = 0; i < list_size; i++) {
         if (list[i]->hit(r, tmin, closest_so_far, temp_rec)) {
